Bug 1242284 - Add local, localOld, and automation Gradle product flavors. r=sebastian
The default configuration is now localDebug for API 21+, and
localOldDebug for <API 21.  This actually paves the way nicely for
automation builds; previously, I was intending to use debug and
release to for local and automation builds, respectively, but this is
more clear.

UltraBlame original commit: cf715a2893acb9a07d74974868301c502f0b5c9d