Bug 1240601 - Set timeout on socket in transport.wait_for_port; r=automatedtester
If the client is able to connect but the server never sends any data,
the socket should time out in order for the function not to hang forever.

That said, polling for ports like this is inherently racy and we should
instead specify and bind what port we wish Marionette to start on in
the future.

UltraBlame original commit: 5c139df9284476d17b5d4f4019ba1fbf12504c07