Bug 1243234 - Move mp4parse into a crate subdirectory. r=kinetik
If we can get this to work it will be less confusing, I think.

Byteorder has to be in a subdir while we're treating it as a mod.

We may have to add yet another directory when we have crate support
but until then all the rust code is at least self-contained..

UltraBlame original commit: 8bc6ad5f9d8f750b8da5643cc18737be2f84ccb4