Bug 1216842 - Part 1: Add null_t into TimingFunction to skip calculation of linear timing function. r=cam
This is a patch for compositor side to represent linear function as null_t/Nothing().

Also the first argument of ToTimingFunction changes to a Maybe<ComputedTimingFunction>.
As a result of this change we can also use ToTimingFunction for animation
effect's timing function.

UltraBlame original commit: 987f0cff946750eb7436c677cd3622a6fdf4d459