Bug 1240796 - Detemplatize getOrCreateSimdTypeDescr(). r=bbouvier
This saves some code size in a cold function, and it makes it possible to pass
in the SIMD type as a dynamic argument.

Also detemplatize the static CreateSimdType() to save some code size.

Replace all of the Get*TypeRepr() self-hosting functions with a single
GetSimdTypeDescr() which takes one of the JS_SIMDTYPEREPR_* constants as an
argument instead.

Total code shrink ~ 32 KB.

UltraBlame original commit: 449c568f3dc59845245086884a3cf56af2aa5fd4