Bug 1194721: Add interface for Gonk sensors daemon, r=gsvelto
This patch adds the interface for the Gonk sensors daemon. There
are only two major interfaces: connect and disconnect. The sensors
daemon is started before connecting and stopped after a disconnect.

Connections are race-free: we first set up a listen socket and let
Gecko listen for connections. Then the daemon gets started with the
socket's address as parameter. When the daemon connects to the
socket, Gecko will continue the startup and signal success to the
caller.

UltraBlame original commit: 15a9b3c68084b86182b555ddbfdba2b1c1b4659c