Bug 1243241 - Make RDTSC monotonic. r=jandem
We assume that the total number of cycles spent executing JS code
during an event is equal to the number of cycles in the "top group",
i.e. a group to which everything belongs. While this is true in
theory, RDTSC is actually non-monotonic, so we can end up with fewer
cycles reported for the top group than for some groups whose execution
was actually shorter. When we end up in this situation, groups with
more cycles than the top group will be reported as using more CPU than
was actually used.

This patch fixes the situation by proxying RDTSC behind a trivial API
that ensures that values are monotonic during each tick.

UltraBlame original commit: 82dc7ec9db670a5e579f2cf0d5ecf54da28485dc