Bug 1219554 - Follow up: clean up memory panel open/destroy; r=jsantell
This commit removes the assignment expression in a return statement thing that
was going on. It is easy to misread and we don't usually use this idiom anywhere
else in the codebase. Additionally, it makes the destroy method clean up some
more members set when opening the tool.

UltraBlame original commit: 61b3d4d0dfd46d91db55c8dbbd41e7652aaa4523