Bug 1236382 - Add commonly used arguments to mach try, remove the extra arguments functionality. r=jgraham
Using nargs='*' in conjunction with nargs=REMAINDER creates an ambiguity when
the argument using nargs='*' is optional, it is not specified, and the user
intends their arguments to be interpreted as "extra" arguments. This commit
removes the nargs=REMAINDER argument for mach_try, and implements its most
common users from trychooser as a set of regular arguments to echo to
generated try syntax.

MozReview-Commit-ID: FOuDZxUfHu9

UltraBlame original commit: c654c659bac1f002d5882892ff3651a7cb1009e6