Bug 1247170 - Part 1: Add "android-gradle-build" Docker image. r=dustin,sebastian
We can't run Docker-in-Docker, so this is a clone of
https://hub.docker.com/r/sonatype/nexus, modified to run additional
commands.  Sonatype Nexus officially supports the Oracle JDK but the
OpenJDK appears to work as well, so we use it.

This bakes a build.sh and Gradle-specific dependencies into the docker
image.  This makes sense in a future where the dependencies this
fetches are baked into the image used to build Fennec.  It makes less
sense right now, when the dependencies are uploaded to tooltool and
subsequently consumed.  It's not critical that we get this right
immediately.

The actual build.sh script runs the Gradle wrapper for a special
dependency project.  The wrapper installs Gradle itself, downloads a
large number of Gradle dependencies, and uses Jake Wharton's
https://github.com/JakeWharton/sdk-manager-plugin/ to download and
install the Android SDK.  (We could use |mach bootstrap|, but that
doesn't (yet) support non-interactive use, and I want to try to use
the SDK Manager more generally.)  These outputs are then packaged for
further use: right now, they're exposed as artifacts, but eventually
we'll upload them directly to tooltool.

MozReview-Commit-ID: 7upsk1ANuxN

UltraBlame original commit: 123c802d928a9a02040cc3316988d9b59db0ea45