Bug 1246115 - Make gSafeBrowsing set the phishing menu item correctly. r=Gijs
Unfortunately, when onLocationChange is fired for an attack site for
the about:blocked error page that we display, content.document has not
been updated with the loaded error document, so
content.document.documentURI will appear to be the previous page that
had been loaded. In this patch, we update the parent's cache of
documentURI in onStateChange as well, since this seems to be fired
after the error page has been loaded.

MozReview-Commit-ID: 1yLAw0JTEC6

UltraBlame original commit: 6ced51b8fdda68356658ab4fea046e2039dffda6