Bug 1244295 - Add junit4 tests for GeckoProfile.getClient & friends. r=mfinkle
Added testGetDir to sanity check how the profile is set up for the test and
left it in as a bonus.

Additionally, changed access levels on the ensureParentDirs method because it
only needed to be `protected` for testing.

MozReview-Commit-ID: CDVQjyf3yP2

UltraBlame original commit: 414d0894875cd433ee42d3ffd5bedbfb4dd6bc4d